Only call `mallinfo` when we know the malloc library supports it
Factor out existing malloc implementation detection code and use
it when emitting malloc arena metrics.

Only dump mallinfo stats if we _know_ that the underlying malloc
implementation supports it. Glibc of course supports it, but we
can't tell for sure that we're running it when observing
`MallocImpl::LibcOrUnknown` because of the "unknown" aspect of it.
We _could_ be running Bob's Artisan Home-Brewed Malloc(tm), which
would be subsumed by `LibcOrUnknown`.

Using a custom (non-glibc) malloc implementation that does _not_
have `mallinfo()`, and then calling it, may trigger internal glibc
lazy malloc heap/arena init code, as glibc erroneously believes
malloc is being bootstrapped by the main thread.

This causes much excitement further down the line.
