+++ searchcore/src/vespa/searchcore/proton/server/proton.cpp
@@ -57,6 +57,7 @@
 #include <vespa/vespalib/net/http/state_server.h>
 #include <vespa/vespalib/util/blockingthreadstackexecutor.h>
 #include <vespa/vespalib/util/cpu_usage.h>
+#include <vespa/vespalib/util/detect_malloc_impl.h>
 #include <vespa/vespalib/util/host_name.h>
 #include <vespa/vespalib/util/lambdatask.h>
 #include <vespa/vespalib/util/mmap_file_allocator_factory.h>
@@ -868,15 +869,26 @@ void Proton::updateMetrics(const metrics::MetricLockGuard&) {
         metrics.resourceUsage.openFileDescriptors.set(FastOS_File::count_open_files());
         metrics.resourceUsage.feedingBlocked.set((usage_filter.acceptWriteOperation() ? 0.0 : 1.0));
 #ifdef __linux__
+        static const auto malloc_impl = vespalib::detect_malloc_impl();
+        // Only dump mallinfo if we _know_ that the underlying malloc implementation
+        // supports it. We can't tell this for sure with LibcOrUnknown because of the
+        // "unknown" aspect of it. It _could_ be Bob's Artisan Home-Brewed Malloc(tm).
+        // Using a custom (non-glibc) malloc implementation that does _not_ have mallinfo()
+        // and then calling it may trigger internal glibc malloc heap/arena init code, as
+        // glibc erroneously believes malloc is being bootstrapped by the main thread.
+        if (malloc_impl == vespalib::MallocImpl::VespaMalloc) {
 #if __GLIBC_PREREQ(2, 33)
-        struct mallinfo2 mallocInfo = mallinfo2();
-        metrics.resourceUsage.mallocArena.set(mallocInfo.arena);
+            struct mallinfo2 mallocInfo = mallinfo2();
+            metrics.resourceUsage.mallocArena.set(mallocInfo.arena);
 #else
-        struct mallinfo mallocInfo = mallinfo();
-        // Vespamalloc reports arena in 1M blocks as an 'int' is too small.
-        // If we use something else than vespamalloc this must be changed.
-        metrics.resourceUsage.mallocArena.set(uint64_t(mallocInfo.arena) * 1_Mi);
+            struct mallinfo mallocInfo = mallinfo();
+            // Vespamalloc reports arena in 1M blocks as an 'int' is too small.
+            // If we use something else than vespamalloc this must be changed.
+            metrics.resourceUsage.mallocArena.set(uint64_t(mallocInfo.arena) * 1_Mi);
 #endif
+        } else {
+            metrics.resourceUsage.mallocArena.set(UINT64_C(0));
+        }
 #else
         metrics.resourceUsage.mallocArena.set(UINT64_C(0));
 #endif
